monitoring/backend/requirements.txt
pandas
psycopg2-binary
requests
